Modified MaterialShowcaseView according to Kanvas UI guidelines. Package refactor.
