+++ library/src/main/java/com/wolox/kanvasshowcaseview/IAnimationFactory.java
@@ -1,0 +1,23 @@
+package com.wolox.kanvasshowcaseview;
+
+import android.graphics.Point;
+import android.view.View;
+
+
+public interface IAnimationFactory {
+
+    void fadeInView(View target, long duration, AnimationStartListener listener);
+
+    void fadeOutView(View target, long duration, AnimationEndListener listener);
+
+    void animateTargetToPoint(KanvasShowcaseView showcaseView, Point point);
+
+    public interface AnimationStartListener {
+        void onAnimationStart();
+    }
+
+    public interface AnimationEndListener {
+        void onAnimationEnd();
+    }
+}
+
