+++ library/src/main/java/com/wolox/kanvasshowcaseview/IDetachedListener.java
@@ -1,0 +1,6 @@
+package com.wolox.kanvasshowcaseview;
+
+
+public interface IDetachedListener {
+    void onShowcaseDetached(KanvasShowcaseView showcaseView, boolean wasDismissed);
+}
