+++ library/src/main/java/com/wolox/kanvasshowcaseview/IShowcaseListener.java
@@ -1,0 +1,7 @@
+package com.wolox.kanvasshowcaseview;
+
+
+public interface IShowcaseListener {
+    void onShowcaseDisplayed(KanvasShowcaseView showcaseView);
+    void onShowcaseDismissed(KanvasShowcaseView showcaseView);
+}
