+++ library/src/main/java/com/wolox/kanvasshowcaseview/shape/Shape.java
@@ -1,0 +1,35 @@
+package com.wolox.kanvasshowcaseview.shape;
+
+import android.graphics.Canvas;
+import android.graphics.Paint;
+
+import com.wolox.kanvasshowcaseview.target.Target;
+
+/**
+ * Specifies a shape of the target (e.g circle, rectangle).
+ * Implementations of this interface will be responsible to draw the shape
+ * at specified center point (x, y).
+ */
+public interface Shape {
+
+    /**
+     * Draw shape on the canvas with the center at (x, y) using Paint object provided.
+     */
+    void draw(Canvas canvas, Paint paint, int x, int y, int padding);
+
+    /**
+     * Get width of the shape.
+     */
+    int getWidth();
+
+    /**
+     * Get height of the shape.
+     */
+    int getHeight();
+
+    /**
+     * Update shape bounds if necessary
+     */
+    void updateTarget(Target target);
+
+}
