--- library/src/main/java/uk/co/deanwild/materialshowcaseview/IAnimationFactory.java
@@ -1,23 +1,0 @@
-package uk.co.deanwild.materialshowcaseview;
-
-import android.graphics.Point;
-import android.view.View;
-
-
-public interface IAnimationFactory {
-
-    void fadeInView(View target, long duration, AnimationStartListener listener);
-
-    void fadeOutView(View target, long duration, AnimationEndListener listener);
-
-    void animateTargetToPoint(MaterialShowcaseView showcaseView, Point point);
-
-    public interface AnimationStartListener {
-        void onAnimationStart();
-    }
-
-    public interface AnimationEndListener {
-        void onAnimationEnd();
-    }
-}
-
