--- library/src/main/java/uk/co/deanwild/materialshowcaseview/IDetachedListener.java
@@ -1,6 +1,0 @@
-package uk.co.deanwild.materialshowcaseview;
-
-
-public interface IDetachedListener {
-    void onShowcaseDetached(MaterialShowcaseView showcaseView, boolean wasDismissed);
-}
