--- library/src/main/java/uk/co/deanwild/materialshowcaseview/IShowcaseListener.java
@@ -1,7 +1,0 @@
-package uk.co.deanwild.materialshowcaseview;
-
-
-public interface IShowcaseListener {
-    void onShowcaseDisplayed(MaterialShowcaseView showcaseView);
-    void onShowcaseDismissed(MaterialShowcaseView showcaseView);
-}
