+++ library/src/main/res/layout/showcase_content.xml
@@ -1,33 +1,32 @@
 <?xml version="1.0" encoding="utf-8"?>
-
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/content_box"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="vertical"
-    android:padding="16dp">
+    android:padding="28dp">
 
     <TextView
         android:id="@+id/tv_content"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:textColor="@android:color/white"
-        android:paddingLeft="5dp"
-        android:textSize="20dp" />
+        android:gravity="end"
+        android:paddingEnd="16dp"
+        android:paddingRight="16dp"
+        android:textSize="26sp" />
 
     <TextView
         android:id="@+id/tv_dismiss"
         android:background="?android:attr/selectableItemBackground"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="20dp"
+        android:layout_gravity="end"
         android:clickable="true"
-        android:paddingBottom="10dp"
-        android:paddingTop="10dp"
-        android:paddingLeft="5dp"
-        android:paddingRight="5dp"
+        android:padding="16dp"
         android:textColor="@android:color/white"
-        android:textSize="22dp"
+        android:textSize="16sp"
+        android:textStyle="bold"
         android:visibility="gone" />
 
 </LinearLayout>
