+++ sample/src/main/java/uk/com/wolox/materialshowcaseviewsample/SequenceExample.java
@@ -1,0 +1,99 @@
+package uk.com.wolox.materialshowcaseviewsample;
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+import android.view.View;
+import android.widget.Button;
+import android.widget.Toast;
+
+import com.wolox.kanvasshowcaseview.KanvasShowcaseSequence;
+import com.wolox.kanvasshowcaseview.KanvasShowcaseView;
+import com.wolox.kanvasshowcaseview.ShowcaseConfig;
+
+
+public class SequenceExample extends AppCompatActivity implements View.OnClickListener {
+
+    private Button mButtonOne;
+    private Button mButtonTwo;
+    private Button mButtonThree;
+
+    private Button mButtonReset;
+
+    private static final String SHOWCASE_ID = "sequence example";
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_sequence_example);
+        mButtonOne = (Button) findViewById(R.id.btn_one);
+        mButtonOne.setOnClickListener(this);
+
+        mButtonTwo = (Button) findViewById(R.id.btn_two);
+        mButtonTwo.setOnClickListener(this);
+
+        mButtonThree = (Button) findViewById(R.id.btn_three);
+        mButtonThree.setOnClickListener(this);
+
+        mButtonReset = (Button) findViewById(R.id.btn_reset);
+        mButtonReset.setOnClickListener(this);
+
+        presentShowcaseSequence(); // one second delay
+    }
+
+    @Override
+    public void onClick(View v) {
+
+        if (v.getId() == R.id.btn_one || v.getId() == R.id.btn_two || v.getId() == R.id.btn_three) {
+
+            presentShowcaseSequence();
+
+        } else if (v.getId() == R.id.btn_reset) {
+
+            KanvasShowcaseView.resetSingleUse(this, SHOWCASE_ID);
+            Toast.makeText(this, "Showcase reset", Toast.LENGTH_SHORT).show();
+        }
+
+    }
+
+    private void presentShowcaseSequence() {
+
+        ShowcaseConfig config = new ShowcaseConfig();
+        config.setDelay(500); // half second between each showcase view
+
+        KanvasShowcaseSequence sequence = new KanvasShowcaseSequence(this, SHOWCASE_ID);
+
+        sequence.setOnItemShownListener(new KanvasShowcaseSequence.OnSequenceItemShownListener() {
+            @Override
+            public void onShow(KanvasShowcaseView itemView, int position) {
+                Toast.makeText(itemView.getContext(), "Item #" + position, Toast.LENGTH_SHORT).show();
+            }
+        });
+
+        sequence.setConfig(config);
+
+        sequence.addSequenceItem(mButtonOne, "This is button one", "GOT IT");
+
+        sequence.addSequenceItem(
+                new KanvasShowcaseView.Builder(this)
+                        .setTarget(mButtonTwo)
+                        .setDismissText("GOT IT")
+                        .setContentText("This is button two")
+                        .withRectangleShape(true)
+                        .build()
+        );
+
+        sequence.addSequenceItem(
+                new KanvasShowcaseView.Builder(this)
+                        .setTarget(mButtonThree)
+                        .setDismissText("GOT IT")
+                        .setContentText("This is button three")
+                        .withRectangleShape()
+                        .build()
+        );
+
+        sequence.start();
+
+    }
+
+}
