+++ sample/src/main/res/layout/activity_custom_example.xml
@@ -4,7 +4,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="uk.co.deanwild.materialshowcaseviewsample.CustomExample">
+    tools:context="uk.com.wolox.materialshowcaseviewsample.CustomExample">
 
     <Button
         android:id="@+id/btn_show"
