+++ sample/src/main/res/layout/activity_sequence_example.xml
@@ -6,7 +6,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="uk.co.deanwild.materialshowcaseviewsample.SequenceExample">
+    tools:context="uk.com.wolox.materialshowcaseviewsample.SequenceExample">
 
     <Button
         android:id="@+id/btn_one"
