+++ sample/src/main/res/layout/activity_simple_single_example.xml
@@ -6,7 +6,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="uk.co.deanwild.materialshowcaseviewsample.SimpleSingleExample">
+    tools:context="uk.com.wolox.materialshowcaseviewsample.SimpleSingleExample">
 
     <Button
         android:id="@+id/btn_show"
